Moved e2e boilerplate to framework
