+++ test/e2e/framework/BUILD
@@ -23,6 +23,7 @@ go_library(
         "rc_util.go",
         "resource_usage_gatherer.go",
         "size.go",
+        "suites.go",
         "test_context.go",
         "util.go",
     ],
@@ -46,6 +47,7 @@ go_library(
         "//pkg/scheduler/nodeinfo:go_default_library",
         "//pkg/util/system:go_default_library",
         "//pkg/util/taints:go_default_library",
+        "//pkg/version:go_default_library",
         "//pkg/volume/util:go_default_library",
         "//staging/src/k8s.io/api/apps/v1:go_default_library",
         "//staging/src/k8s.io/api/core/v1:go_default_library",
@@ -98,6 +100,7 @@ go_library(
         "//test/e2e/framework/resource:go_default_library",
         "//test/e2e/framework/ssh:go_default_library",
         "//test/e2e/framework/testfiles:go_default_library",
+        "//test/e2e/manifest:go_default_library",
         "//test/utils:go_default_library",
         "//test/utils/image:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
